Refactor: Move getShelfData logic to UserRepository
