+++ app/src/main/java/org/ole/planet/myplanet/repository/UserRepository.kt
@@ -109,4 +109,5 @@ interface UserRepository {
     suspend fun markAchievementUploaded(id: String, rev: String?)
     fun bulkInsertAchievementsFromSync(realm: io.realm.Realm, jsonArray: com.google.gson.JsonArray)
     fun bulkInsertUsersFromSync(realm: io.realm.Realm, jsonArray: com.google.gson.JsonArray, settings: android.content.SharedPreferences)
+    suspend fun getShelfData(userId: String?, jsonDoc: JsonObject?, myLibs: JsonArray, myCourseIds: JsonArray): JsonObject
 }
